+++ src/gufo/acme/clients/__init__.py
@@ -9,5 +9,6 @@ Various ACME clients implementations.
 
 The following clients are provided out-of-the-box:
 
-* [base
+* [base][gufo.acme.clients.base] - Base class
+* [dav][gufo.acme.clients.dav] - http-01 WebDAV fulfillment.
 """